pyproject.toml
[project]
name = "lifeos"
version = "0.1.0"
description = "Personal Telegram assistant with LLM-powered task management"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-telegram-bot>=21.0",
    "openai>=1.0",
    "apscheduler>=3.10",
    "python-dotenv>=1.0",
]

[project.scripts]
lifeos = "lifeos.__main__:main"
lifeos-cli = "lifeos.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lifeos"]

[dependency-groups]
dev = [
    "ty>=0.0.9",
]
